lib/assets/html/backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-dotenv==1.0.1
PyJWT==2.8.0
Werkzeug==3.0.1
SQLAlchemy==2.0.27
python-engineio==4.9.0
python-socketio==5.11.0
mysqlclient==2.2.7
PyMySQL==1.1.1
eventlet==0.39.1
